Some includes can be used on Mac OS X now (bug 3515, r=ds).

+++ public/tier0/memalloc.h
@@ -117,7 +117,7 @@ inline void *MemAlloc_AllocAligned( size_t size, size_t align )
 {
 	unsigned char *pAlloc, *pResult;
 
-	if (!IsPowerOfTwo(align))
+	if (!IsPowerOfTwo(uint(align)))
 		return NULL;
 
 	align = (align > sizeof(void *) ? align : sizeof(void *)) - 1;
@@ -135,7 +135,7 @@ inline void *MemAlloc_AllocAligned( size_t size, size_t align, const char *pszFi
 {
 	unsigned char *pAlloc, *pResult;
 
-	if (!IsPowerOfTwo(align))
+	if (!IsPowerOfTwo(uint(align)))
 		return NULL;
 
 	align = (align > sizeof(void *) ? align : sizeof(void *)) - 1;
@@ -151,7 +151,7 @@ inline void *MemAlloc_AllocAligned( size_t size, size_t align, const char *pszFi
 
 inline void *MemAlloc_ReallocAligned( void *ptr, size_t size, size_t align )
 {
-	if ( !IsPowerOfTwo( align ) )
+	if ( !IsPowerOfTwo( uint(align) ) )
 		return NULL;
 
 	// Don't change alignment between allocation + reallocation.
